+++ back/src/bspaoh/resolvers/queries/__tests__/bspaohpdf.integration.ts
@@ -1,38 +1,41 @@
 import { resetDatabase } from "../../../../../integration-tests/helper";
 import {
-  companyFactory,
   userWithCompanyFactory,
   userWithAccessTokenFactory
 } from "../../../../__tests__/factories";
 import makeClient from "../../../../__tests__/testClient";
 import { ErrorCode } from "../../../../common/errors";
-import { bsdaFactory } from "../../../__tests__/factories";
+import { bspaohFactory } from "../../../__tests__/factories";
 import { Query } from "../../../../generated/graphql/types";
+
 import { GovernmentPermission } from "@prisma/client";
-import supertest from "supertest";
 import { faker } from "@faker-js/faker";
 import { app } from "../../../../server";
 import { gql } from "graphql-tag";
+import supertest from "supertest";
 
-const BSDA_PDF = gql`
-  query BsdaPdf($id: ID!) {
-    bsdaPdf(id: $id) {
+const BSPAOH_PDF = gql`
+  query Bspaohdf($id: ID!) {
+    bspaohPdf(id: $id) {
       token
     }
   }
 `;
 
-describe("Query.BsdaPdf", () => {
+describe("Query.BspaohPdf", () => {
   afterEach(resetDatabase);
 
   it("should disallow unauthenticated user", async () => {
     const { query } = makeClient();
-    const bsda = await bsdaFactory({
-      opt: {}
+    const { company } = await userWithCompanyFactory("MEMBER");
+    const paoh = await bspaohFactory({
+      opt: {
+        emitterCompanySiret: company.siret
+      }
     });
 
-    const { errors } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
+    const { errors } = await query<Pick<Query, "bspaohPdf">>(BSPAOH_PDF, {
+      variables: { id: paoh.id }
     });
     expect(errors).toEqual([
       expect.objectContaining({
@@ -44,20 +47,23 @@ describe("Query.BsdaPdf", () => {
     ]);
   });
 
-  it("should forbid access to user not on the bsd (simple bsda)", async () => {
-    const bsda = await bsdaFactory({
-      opt: {}
+  it("should forbid access to user not on the bsd", async () => {
+    const { company } = await userWithCompanyFactory("MEMBER");
+    const paoh = await bspaohFactory({
+      opt: {
+        emitterCompanySiret: company.siret
+      }
     });
-    const { user } = await userWithCompanyFactory("MEMBER");
+    const { user: otherUser } = await userWithCompanyFactory("MEMBER");
 
-    const { query } = makeClient(user);
-    const { errors } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
+    const { query } = makeClient(otherUser);
+    const { errors } = await query<Pick<Query, "bspaohPdf">>(BSPAOH_PDF, {
+      variables: { id: paoh.id }
     });
     expect(errors).toEqual([
       expect.objectContaining({
         message:
-          "Vous n'êtes pas autorisé à accéder au récépissé PDF de ce BSDA.",
+          "Vous n'êtes pas autorisé à accéder au récépissé PDF de ce BSPAOH.",
         extensions: expect.objectContaining({
           code: ErrorCode.FORBIDDEN
         })
@@ -67,7 +73,7 @@ describe("Query.BsdaPdf", () => {
 
   it("should return a token for requested id", async () => {
     const { user, company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
+    const paoh = await bspaohFactory({
       opt: {
         emitterCompanySiret: company.siret
       }
@@ -75,62 +81,16 @@ describe("Query.BsdaPdf", () => {
 
     const { query } = makeClient(user);
 
-    const { data } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
+    const { data } = await query<Pick<Query, "bspaohPdf">>(BSPAOH_PDF, {
+      variables: { id: paoh.id }
     });
 
-    expect(data.bsdaPdf.token).toBeTruthy();
-  });
-
-  it("should return a token for requested id if current user is not on the bsda but on a parent bsda", async () => {
-    const { user, company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: company.siret
-      }
-    });
-
-    const forwardingBsda = await bsdaFactory({
-      opt: {
-        forwarding: { connect: { id: bsda.id } }
-      }
-    });
-
-    const { query } = makeClient(user);
-
-    const { data } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: forwardingBsda.id }
-    });
-
-    expect(data.bsdaPdf.token).toBeTruthy();
-  });
-
-  it("should return a token for requested id if current user is an intermediary", async () => {
-    const otherCompany = await companyFactory();
-    const { user, company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: otherCompany.siret,
-        intermediaries: {
-          create: [
-            { siret: company.siret!, name: company.name, contact: "joe" }
-          ]
-        }
-      }
-    });
-
-    const { query } = makeClient(user);
-
-    const { data } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
-    });
-
-    expect(data.bsdaPdf.token).toBeTruthy();
+    expect(data.bspaohPdf.token).toBeTruthy();
   });
 
   it("should allow pdf access to user authenticated with a token when tied to a government account with relevant perms", async () => {
     const { company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
+    const paoh = await bspaohFactory({
       opt: {
         emitterCompanySiret: company.siret
       }
@@ -154,18 +114,19 @@ describe("Query.BsdaPdf", () => {
     const res = await request
       .post("/")
       .send({
-        query: `{bsdaPdf(id: "${bsda.id}") {token}}`
+        query: `{bspaohPdf(id: "${paoh.id}") {token}}`
       })
       .set("Authorization", `Bearer ${accessToken}`)
       .set("X-Forwarded-For", allowedIP);
     const { errors, data } = res.body;
 
     expect(errors).toBeUndefined();
-    expect(data.bsdaPdf.token).toBeTruthy();
+    expect(data.bspaohPdf.token).toBeTruthy();
   });
+
   it("should forbid pdf access to user authenticated with a token when tied to a government account without relevant perms", async () => {
     const { company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
+    const paoh = await bspaohFactory({
       opt: {
         emitterCompanySiret: company.siret
       }
@@ -188,7 +149,7 @@ describe("Query.BsdaPdf", () => {
     const res = await request
       .post("/")
       .send({
-        query: `{bsdaPdf(id: "${bsda.id}") {token}}`
+        query: `{bspaohPdf(id: "${paoh.id}") {token}}`
       })
       .set("Authorization", `Bearer ${accessToken}`)
       .set("X-Forwarded-For", allowedIP);
@@ -216,7 +177,7 @@ describe("Query.BsdaPdf", () => {
         }
       }
     });
-    const bsda = await bsdaFactory({
+    const paoh = await bspaohFactory({
       opt: {
         emitterCompanySiret: company.siret
       }
@@ -225,7 +186,7 @@ describe("Query.BsdaPdf", () => {
     const res = await request
       .post("/")
       .send({
-        query: `{bsdaPdf(id: "${bsda.id}") {token}}`
+        query: `{bspaohPdf(id: "${paoh.id}") {token}}`
       })
       .set("Authorization", `Bearer ${accessToken}`)
       .set("X-Forwarded-For", userIP); // IPs do not match