+++ back/src/forms/resolvers/queries/__tests__/formPdf.integration.ts
@@ -1,38 +1,42 @@
 import { resetDatabase } from "../../../../../integration-tests/helper";
 import {
-  companyFactory,
   userWithCompanyFactory,
-  userWithAccessTokenFactory
+  userWithAccessTokenFactory,
+  formFactory
 } from "../../../../__tests__/factories";
+import { GovernmentPermission } from "@prisma/client";
+
 import makeClient from "../../../../__tests__/testClient";
 import { ErrorCode } from "../../../../common/errors";
-import { bsdaFactory } from "../../../__tests__/factories";
+
 import { Query } from "../../../../generated/graphql/types";
-import { GovernmentPermission } from "@prisma/client";
 import supertest from "supertest";
 import { faker } from "@faker-js/faker";
 import { app } from "../../../../server";
+
 import { gql } from "graphql-tag";
 
-const BSDA_PDF = gql`
-  query BsdaPdf($id: ID!) {
-    bsdaPdf(id: $id) {
+const FORM_PDF = gql`
+  query FormPdf($id: ID!) {
+    formPdf(id: $id) {
       token
     }
   }
 `;
 
-describe("Query.BsdaPdf", () => {
+describe("Query.FormPdf", () => {
   afterEach(resetDatabase);
 
   it("should disallow unauthenticated user", async () => {
     const { query } = makeClient();
-    const bsda = await bsdaFactory({
-      opt: {}
+    const { user, company } = await userWithCompanyFactory("MEMBER");
+    const bsdd = await formFactory({
+      ownerId: user.id,
+      opt: { emitterCompanySiret: company.siret }
     });
 
-    const { errors } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
+    const { errors } = await query<Pick<Query, "formPdf">>(FORM_PDF, {
+      variables: { id: bsdd.id }
     });
     expect(errors).toEqual([
       expect.objectContaining({
@@ -44,20 +48,21 @@ describe("Query.BsdaPdf", () => {
     ]);
   });
 
-  it("should forbid access to user not on the bsd (simple bsda)", async () => {
-    const bsda = await bsdaFactory({
-      opt: {}
+  it("should forbid access to user not on the bsdd", async () => {
+    const { user, company } = await userWithCompanyFactory("MEMBER");
+    const bsdd = await formFactory({
+      ownerId: user.id,
+      opt: { emitterCompanySiret: company.siret }
     });
-    const { user } = await userWithCompanyFactory("MEMBER");
+    const { user: otherUser } = await userWithCompanyFactory("MEMBER");
 
-    const { query } = makeClient(user);
-    const { errors } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
+    const { query } = makeClient(otherUser);
+    const { errors } = await query<Pick<Query, "formPdf">>(FORM_PDF, {
+      variables: { id: bsdd.id }
     });
     expect(errors).toEqual([
       expect.objectContaining({
-        message:
-          "Vous n'êtes pas autorisé à accéder au récépissé PDF de ce BSDA.",
+        message: "Vous n'êtes pas autorisé à accéder à ce bordereau",
         extensions: expect.objectContaining({
           code: ErrorCode.FORBIDDEN
         })
@@ -67,73 +72,25 @@ describe("Query.BsdaPdf", () => {
 
   it("should return a token for requested id", async () => {
     const { user, company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: company.siret
-      }
+    const bsdd = await formFactory({
+      ownerId: user.id,
+      opt: { emitterCompanySiret: company.siret }
     });
 
     const { query } = makeClient(user);
 
-    const { data } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
+    const { data } = await query<Pick<Query, "formPdf">>(FORM_PDF, {
+      variables: { id: bsdd.id }
     });
 
-    expect(data.bsdaPdf.token).toBeTruthy();
-  });
-
-  it("should return a token for requested id if current user is not on the bsda but on a parent bsda", async () => {
-    const { user, company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: company.siret
-      }
-    });
-
-    const forwardingBsda = await bsdaFactory({
-      opt: {
-        forwarding: { connect: { id: bsda.id } }
-      }
-    });
-
-    const { query } = makeClient(user);
-
-    const { data } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: forwardingBsda.id }
-    });
-
-    expect(data.bsdaPdf.token).toBeTruthy();
-  });
-
-  it("should return a token for requested id if current user is an intermediary", async () => {
-    const otherCompany = await companyFactory();
-    const { user, company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: otherCompany.siret,
-        intermediaries: {
-          create: [
-            { siret: company.siret!, name: company.name, contact: "joe" }
-          ]
-        }
-      }
-    });
-
-    const { query } = makeClient(user);
-
-    const { data } = await query<Pick<Query, "bsdaPdf">>(BSDA_PDF, {
-      variables: { id: bsda.id }
-    });
-
-    expect(data.bsdaPdf.token).toBeTruthy();
+    expect(data.formPdf.token).toBeTruthy();
   });
 
   it("should allow pdf access to user authenticated with a token when tied to a government account with relevant perms", async () => {
-    const { company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: company.siret
-      }
+    const { user, company } = await userWithCompanyFactory("MEMBER");
+    const bsdd = await formFactory({
+      ownerId: user.id,
+      opt: { emitterCompanySiret: company.siret }
     });
 
     const allowedIP = faker.internet.ipv4();
@@ -154,22 +111,23 @@ describe("Query.BsdaPdf", () => {
     const res = await request
       .post("/")
       .send({
-        query: `{bsdaPdf(id: "${bsda.id}") {token}}`
+        query: `{formPdf(id: "${bsdd.id}") {token}}`
       })
       .set("Authorization", `Bearer ${accessToken}`)
       .set("X-Forwarded-For", allowedIP);
     const { errors, data } = res.body;
 
     expect(errors).toBeUndefined();
-    expect(data.bsdaPdf.token).toBeTruthy();
+    expect(data.formPdf.token).toBeTruthy();
   });
+
   it("should forbid pdf access to user authenticated with a token when tied to a government account without relevant perms", async () => {
-    const { company } = await userWithCompanyFactory("MEMBER");
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: company.siret
-      }
+    const { user, company } = await userWithCompanyFactory("MEMBER");
+    const bsdd = await formFactory({
+      ownerId: user.id,
+      opt: { emitterCompanySiret: company.siret }
     });
+
     const allowedIP = faker.internet.ipv4();
 
     const { accessToken } = await userWithAccessTokenFactory({
@@ -188,7 +146,7 @@ describe("Query.BsdaPdf", () => {
     const res = await request
       .post("/")
       .send({
-        query: `{bsdaPdf(id: "${bsda.id}") {token}}`
+        query: `{formPdf(id: "${bsdd.id}") {token}}`
       })
       .set("Authorization", `Bearer ${accessToken}`)
       .set("X-Forwarded-For", allowedIP);
@@ -200,7 +158,7 @@ describe("Query.BsdaPdf", () => {
   });
 
   it("should forbid user authenticated with a token tied to a government account when IPs do not match", async () => {
-    const { company } = await userWithCompanyFactory("MEMBER");
+    const { user, company } = await userWithCompanyFactory("MEMBER");
     const request = supertest(app);
 
     const allowedIP = faker.internet.ipv4();
@@ -216,16 +174,15 @@ describe("Query.BsdaPdf", () => {
         }
       }
     });
-    const bsda = await bsdaFactory({
-      opt: {
-        emitterCompanySiret: company.siret
-      }
+    const bsdd = await formFactory({
+      ownerId: user.id,
+      opt: { emitterCompanySiret: company.siret }
     });
 
     const res = await request
       .post("/")
       .send({
-        query: `{bsdaPdf(id: "${bsda.id}") {token}}`
+        query: `{formPdf(id: "${bsdd.id}") {token}}`
       })
       .set("Authorization", `Bearer ${accessToken}`)
       .set("X-Forwarded-For", userIP); // IPs do not match